% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepost_mi_list.R
\name{plot.prepost_mi}
\alias{plot.prepost_mi}
\title{Plot Results from Multiply Imputed Resilience Analysis}
\usage{
\method{plot}{prepost_mi}(x, type = "coefficients", ...)
}
\arguments{
\item{x}{An object of class \code{prepost_resilience_mi_list}.}

\item{type}{Type of plot: \code{"estimates"} (default) or \code{"fmi"}.}

\item{...}{Additional arguments passed to plotting functions.}
}
\description{
Plot method for visualizing imputation diagnostics
}
