% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepost_mi_list.R
\name{print.prepost_resilience_mi_list}
\alias{print.prepost_resilience_mi_list}
\title{Print Results from Multiply Imputed Resilience Analysis}
\usage{
\method{print}{prepost_resilience_mi_list}(x, ...)
}
\arguments{
\item{x}{An object of class \code{prepost_resilience_mi_list}.}

\item{...}{Additional arguments passed to plotting functions.}
}
\description{
S3 print method for prepost_resilience_mi_list
}
