\name{resilience-package}
\alias{resilience-package}
\docType{package}
\title{
Identifying Predictors of Resilience to Stressors in Single-Arm Studies of Pre-Post Change}
\description{
Studies of resilience in older adults are typically conducted with a single-arm where everyone experiences the stressor. The simplistic
approach of regressing change versus baseline yields biased estimates due to mathematical coupling and regression-to-the-mean. This function provides a method to correct the bias.
}
\details{
\tabular{ll}{
Package: \tab resilience\cr
Type: \tab Package\cr
Version: \tab 2023-06.18\cr
Date: \tab 2023-06.18\cr
License: \tab GPL (version 2)\cr
LazyLoad: \tab yes\cr
}
The package contains a singe function \code{prepost}, which implements the method of Varadhan, Zhu, and Bandeen-Roche (2023) to correct for the bias of naive analysis of pre-post change in studies of resilience.
}
\author{
Ravi Varadhan.

Maintainers: Ravi Varadhan <rvaradhan@jhmi.edu>
}
\references{
 Varadhan, R., Zhu, J., and Bandeen-Roche, K (2023). Identifying Predictors of Resilience to Stressors
in Single-Arm Studies of Pre-Post Change. \emph{Biostatistics}. To appear. 
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%~~ the R documentation directory ~~
\keyword{package}
