% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resilience.R
\name{resindex}
\alias{resindex}
\title{Resilience indices}
\usage{
resindex(data, event, ntemp, index)
}
\arguments{
\item{data}{A \link{matrix} or \link{data.frame} with two columns temp (Date) and performance (Numeric)}

\item{event}{Date. Vector of dates of the disturbance events}

\item{ntemp}{Double. Number of temporal steps to include in the calculations}

\item{index}{Character. Indices to be calculated ('resilience', 'resistance' or 'recovery').
Various elements can be concatenated in the vector.}
}
\value{
A \link{data.frame}
}
\description{
Resilience indices
}
