% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepost_mi_list.R
\name{summary.prepost_resilience_mi_list}
\alias{summary.prepost_resilience_mi_list}
\title{Summary Results from Multiply Imputed Resilience Analysis}
\usage{
\method{summary}{prepost_resilience_mi_list}(object, ...)
}
\arguments{
\item{object}{An object of class \code{prepost_resilience_mi_list}.}

\item{...}{Additional arguments passed to plotting functions.}
}
\description{
S3 summary method for prepost_resilience_mi_list
}
