% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resilience-package.R
\docType{package}
\name{resilience-package}
\alias{resilience}
\alias{resilience-package}
\title{resilience: Calculate resilience indices for ecological systems}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EcologyR/resilience}
  \item Report bugs at \url{https://github.com/EcologyR/resilience/issues}
}

}
\author{
\strong{Maintainer}: Verónica Cruz-Alonso \email{veronica.cral@gmail.com} (\href{https://orcid.org/0000-0002-0642-036X}{ORCID})

Authors:
\itemize{
  \item Julen Astigarraga (\href{https://orcid.org/0000-0001-9520-3713}{ORCID})
  \item Francisco Rodriguez-Sanchez \email{f.rodriguez.sanc@gmail.com} (\href{https://orcid.org/0000-0002-7981-1599}{ORCID})
  \item Enrique Andivia (\href{https://orcid.org/0000-0002-9096-3294}{ORCID})
}

}
\keyword{internal}
