% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tkr}
\alias{tkr.dat}
\title{Pre-post stressor response data}
\format{
An object of class \code{data.frame} with 900 rows and 6 columns.
}
\usage{
data(tkr)
}
\description{
A data set with 900 rows and 6 columns
\describe{
  \item{gender}{Binary variable, 555 Female and 345 Male}
  \item{age}{Continuous variable, range from 60 to 100}
  \item{bmi}{Continuous variable, range from 16 to 55, with 44 missing values}
  \item{smoker}{Three-level factor, 819 no, 25 yes, and 56 unknown}
  \item{post.Y}{Continuous variable, post-stressor measure, range from 10 to 58, with 43 missing values}
  \item{pre.Y}{Continuous variable, pre-stressor measure, range from 9 to 66, with 196 missing values}
}
}
\keyword{datasets}
